+++ models.ts
@@ -1,0 +1,11 @@
+export const enum Commands {
+  mouseUp = 'mouse_up',
+  mouseDown = 'mouse_down',
+  mouseLeft = 'mouse_left',
+  mouseRight = 'mouse_right',
+  mousePosition = 'mouse_position',
+  drawCircle = 'draw_circle',
+  drawRectangle = 'draw_rectangle',
+  drawSquare = 'draw_square',
+  prntScrn = 'prnt_scrn',
+}
